test/requirements.txt
pytest>=8.0.0
boto3>=1.40.0
moto>=5.0.0
pillow>=11.0.0
requests>=2.32.0
pytest-cov>=4.0.0
pytest-mock>=3.0.0